Draw: headless draw final milestone reached
git-svn-id: svn://ultimatepp.org/upp/trunk@1436 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/ide/Browser/Item.cpp
@@ -61,7 +61,7 @@ static bool sOperatorTab[256];
 
 INITBLOCK {
 	for(const char *s = "!+-*^/%~&|=[]:?."; *s; s++)
-		sOperatorTab[*s] = true;
+		sOperatorTab[(int)*s] = true;
 }
 
 inline bool sOperator(byte c)
